ARAMBH_99_RAG_AI_MODEL/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
huggingface-hub
gradio
python-dotenv
chromadb
sqlite3
